fix: remove old CodecStats extension, pow->double cast

+++ openchat-flutter/lib/core/audio/epc_codec.dart
@@ -302,7 +302,7 @@ class EpcCodec {
       for (int eo = off + 5; eo < off + 5 + dataLen; eo += 12) {
         var tag = EpcTag.unpack(epcData.sublist(eo, eo + 12));
         if (tag.onsetFlag == 2) { active.remove(tag.trackId); continue; }
-        double freq = 440 * pow(2, (tag.midiNote + tag.cent / 100 - 69) / 12);
+        double freq = 440 * pow(2, (tag.midiNote + tag.cent / 100 - 69) / 12) as double;
         if (tag.rms > 0) {
           active[tag.trackId] = _SynthTone(
             freq: freq, harmonics: tag.harmonics, rms: tag.rms, velocity: tag.velocity,
